new/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
schedule==1.2.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.4.0
python-dotenv==1.0.0
